+++ app/Http/Controllers/backend/ClientController.php
@@ -3,12 +3,15 @@
 namespace App\Http\Controllers\backend;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Models\Client;
 
 class ClientController extends Controller
 {
     public function ClientView(){
+
+        abort_if(Gate::denies('client.view'), 403);
 
         $data['allData'] = Client::all();
 
@@ -17,6 +20,8 @@ class ClientController extends Controller
     }
 
     public function ClientAdd(){
+
+        abort_if(Gate::denies('clients.add'), 403);
 
         return view('backend.client.add_client');
 
@@ -54,6 +59,8 @@ class ClientController extends Controller
 
     public function ClientEdit($id){
 
+        abort_if(Gate::denies('clients.edit'), 403);
+
         $editData = Client::find($id);
         return view('backend.client.edit_client',compact('editData'));
 
@@ -83,6 +90,8 @@ class ClientController extends Controller
     }
 
     public function ClientDelete($id){
+
+        abort_if(Gate::denies('clients.delete'), 403);
 
         $client = Client::find($id);
         $client->delete();
